my_project/requirements.txt
Flask>=2.2.0
Pillow>=8.3.0
PyYAML>=6.0
matplotlib>=3.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
seaborn>=0.11.0
tensorflow>=2.10.0
